api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pytz==2024.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.1
python-multipart==0.0.9
requests==2.32.3
Jinja2==3.1.4
